freelance/backend/requirements.txt
fastapi==0.109.0
pydantic==2.6.0
python-dotenv==1.0.1
python_bcrypt==0.3.2
SQLAlchemy==2.0.25
psycopg2-binary
uvicorn
gunicorn